# docs/api/swagger.yaml
components:
  schemas:
...
    ApiError:
      properties:
        code:
          example: BAD_REQUEST
          type: string
        details:
          description: |-
            Details contains additional error context (field names, validation rules, etc.)
            Type varies by error code: map[string]string for validation, string for general errors
          example: '{"field":"email","rule":"required"}'
        message:
          example: Invalid request parameters
          type: string
      type: object
...
    AuditLog:
      properties:
        actor:
          type: string
        actorRole:
          type: string
        duration:
          description: milliseconds
          type: integer
        error:
          type: string
        id:
          type: string
        issueNumber:
          type: integer
        operation:
          type: string
        output:
          type: string
        repository:
          type: string
        resourceId:
          type: string
        resourceType:
          type: string
        result:
          type: string
        sessionId:
          type: string
        timestamp:
          type: string
      type: object
    AuditSummary:
      properties:
        averageDurationMs:
          type: integer
        byOperation:
          additionalProperties:
            type: integer
          type: object
        byResult:
          additionalProperties:
            type: integer
          type: object
        firstTimestamp:
          type: string
        lastTimestamp:
          type: string
        sessionId:
          type: string
        totalCount:
          type: integer
        totalDurationMs:
          type: integer
      type: object
...
    PaginatedResponse-AuditLog:
      properties:
        data:
          items:
            $ref: '#/components/schemas/AuditLog'
          type: array
          uniqueItems: false
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
...
    Pagination:
      properties:
        page:
          example: 1
          type: integer
        pageSize:
          example: 20
          type: integer
        totalItems:
          example: 100
          type: integer
        totalPages:
          example: 5
          type: integer
      type: object
...
info:
  description: Automated development agent system - from GitHub Issue to Pull Request
  title: Litchi API
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/v1/audit:
    get:
      description: Retrieve audit logs with filtering by session, repository, actor,
        operation, result, and time range. Supports pagination.
      parameters:
      - description: 'Page number (default: 1)'
        example: 1
        in: query
        name: page
        schema:
          type: integer
      - description: 'Page size (default: 50, max: 100)'
        example: 50
        in: query
        name: pageSize
        schema:
          type: integer
      - description: Filter by session ID
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: query
        name: sessionId
        schema:
          type: string
      - description: Filter by repository (owner/repo)
        example: '"owner/repo"'
        in: query
        name: repository
        schema:
          type: string
      - description: Filter by actor username
        example: '"username"'
        in: query
        name: actor
        schema:
          type: string
      - description: Filter by operation type
        example: '"session_start"'
        in: query
        name: operation
        schema:
          type: string
      - description: Filter by result (success/failed/denied)
        example: '"success"'
        in: query
        name: result
        schema:
          type: string
      - description: Filter by start time (RFC3339)
        example: '"2024-01-01T00:00:00Z"'
        in: query
        name: startTime
        schema:
          type: string
      - description: Filter by end time (RFC3339)
        example: '"2024-01-02T00:00:00Z"'
        in: query
        name: endTime
        schema:
          type: string
      - description: 'Order by field (default: timestamp desc)'
        example: '"timestamp desc"'
        in: query
        name: orderBy
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse-AuditLog'
          description: Audit logs retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: List audit logs
      tags:
      - audit
  /api/v1/audit/{id}:
    get:
      description: Retrieve a single audit log entry by its unique identifier.
      parameters:
      - description: Audit log ID (UUID format)
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLog'
          description: Audit log retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid audit log ID format
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Audit log not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Get audit log
      tags:
      - audit
  /api/v1/audit/repositories/{repository}:
    get:
      description: Retrieve all audit logs for a specific repository (owner/repo format).
        Supports pagination.
      parameters:
      - description: Repository name (owner/repo)
        example: '"owner/repo"'
        in: path
        name: repository
        required: true
        schema:
          type: string
      - description: 'Page number (default: 1)'
        example: 1
        in: query
        name: page
        schema:
          type: integer
      - description: 'Page size (default: 50)'
        example: 50
        in: query
        name: pageSize
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse-AuditLog'
          description: Audit logs retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid repository name or query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: List audit logs by repository
      tags:
      - audit
  /api/v1/audit/sessions/{sessionId}:
    get:
      description: Retrieve all audit logs associated with a specific work session.
        Supports pagination.
      parameters:
      - description: Session ID (UUID format)
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: path
        name: sessionId
        required: true
        schema:
          type: string
      - description: 'Page number (default: 1)'
        example: 1
        in: query
        name: page
        schema:
          type: integer
      - description: 'Page size (default: 50)'
        example: 50
        in: query
        name: pageSize
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse-AuditLog'
          description: Audit logs retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid session ID or query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: List audit logs by session
      tags:
      - audit
  /api/v1/audit/sessions/{sessionId}/summary:
    get:
      description: Retrieve aggregated statistics of audit logs for a specific work
        session, including counts by operation type and result status.
      parameters:
      - description: Session ID (UUID format)
        example: '"550e8400-e29b-41d4-a716-446655440000"'
        in: path
        name: sessionId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditSummary'
          description: Audit summary retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid session ID format
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Get session audit summary
      tags:
      - audit
...